simplify entities, update API to use them, change response
Bothan’s model’s are flat and only share a name identifier. There is no
relation between the /metrics returned by Bothan and the Metric model.
As such the use of multiple entities to present the API content - as
detailed in this commit - is merited
This change specifies that a metric name should not be returned on a
GET, so is no longer exposed