Large reorganisation of API, see commit msg
Logic reshuffled for use with Swagger-UI.
There is now effectively one namespace - metrics. The <metric>/metadata
endpoint is kept for now to illustrate differences in grape-entity vs
vanilla grape in terms of how documentation bubbles up to swagger-ui.
The majority of methods/blocks are nested under :metrics to keep the
:metric param description is DRY